fix: removed old async await code
