+++ Sources/Asset.swift
@@ -246,27 +246,6 @@ extension Asset: ResourceQueryable {
                             }
         })
     }
-    
-    public func asyncFetch<ResourceType>(_ completion: @escaping (Result<ResourceType, Error>, ResponseType) -> Void) async
-        where ResourceType: EndpointAccessible, ResourceType: Decodable {
-        guard let uid = self.uid else { fatalError("Please provide Asset uid") }
-        await self.stack.asyncFetch(endpoint: ResourceType.endpoint,
-                         cachePolicy: self.cachePolicy,
-                         parameters: parameters + [QueryParameter.uid: uid],
-                         headers: headers,
-                         then: { (result: Result<ContentstackResponse<ResourceType>, Error>, response: ResponseType) in
-                            switch result {
-                            case .success(let contentStackResponse):
-                                if let resource = contentStackResponse.items.first {
-                                    completion(.success(resource), response)
-                                } else {
-                                    completion(.failure(SDKError.invalidUID(string: uid)), response)
-                                }
-                            case .failure(let error):
-                                completion(.failure(error), response)
-                            }
-        })
-    }
 
     public func fetch<ResourceType>() async throws -> (Result<ResourceType, Error>, ResponseType)
         where ResourceType: EndpointAccessible, ResourceType: Decodable {
