+++ Sources/Stack.swift
@@ -190,33 +190,6 @@ public class Stack: CachePolicyAccessible {
             }
         })
     }
-    
-    internal func asyncFetch<ResourceType>(endpoint: Endpoint,
-                                           cachePolicy: CachePolicy,
-                                           parameters: Parameters = [:],
-                                           headers: [String: String] = [:])
-        async throws -> (Result<ResourceType, Error>, ResponseType)
-        where ResourceType: Decodable {
-            
-            let url = self.url(endpoint: endpoint, parameters: parameters)
-            
-            do {
-                let (result, responseType): (Result<Data, Error>, ResponseType) = try await self.asyncFetchUrl(url: url, headers: headers, cachePolicy: cachePolicy)
-                switch result {
-                case .success(let data):
-                    do {
-                        let jsonParse = try self.jsonDecoder.decode(ResourceType.self, from: data)
-                        return (.success(jsonParse), responseType)
-                    } catch {
-                        throw error
-                    }
-                case .failure(let error):
-                    return (.failure(error), responseType)
-                }
-            } catch {
-                throw error
-            }
-    }
 
     private func fetchUrl(_ url: URL, headers:[String: String], cachePolicy: CachePolicy, then completion: @escaping ResultsHandler<Data>) {
         var dataTask: URLSessionDataTask?
@@ -275,15 +248,6 @@ public class Stack: CachePolicyAccessible {
         performDataTask(dataTask!, request: request, cachePolicy: cachePolicy, then: completion)
     }
     
-    private func asyncFetchUrl(url: URL, headers: [String: String], cachePolicy: CachePolicy) async throws -> (Result<Data, Error>, ResponseType) {
-        do {
-            let (data, response): (Data, ResponseType) = try await fetchDataAsync(url: url, headers: headers)
-            return (.success(data), response)
-        } catch {
-            return (.failure(error), .network)
-        }
-    }
-    
     private func fetchDataAsync(url: URL, headers: [String: String]) async throws -> (Data, ResponseType) {
         return try await withCheckedThrowingContinuation { continuation in
             self.fetchUrl(url, headers: headers, cachePolicy: cachePolicy, then: { (result: Result<Data, Error>, responseType: ResponseType) in
@@ -337,6 +301,7 @@ public class Stack: CachePolicyAccessible {
         }
         completion(Result.failure(SDKError.cacheError), .cache)
     }
+
     private func fulfillRequestWithCache(_ request: URLRequest) async throws -> (Result<Data, Error>, ResponseType) {
         if let data = self.cachedResponse(for: request) {
             return (.success(data), .cache)
@@ -355,89 +320,8 @@ public class Stack: CachePolicyAccessible {
         }
         return nil
     }
-    
-    internal func asyncFetch<ResourceType>(endpoint: Endpoint,
-                                      cachePolicy: CachePolicy,
-                                      parameters: Parameters = [:],
-                                      headers: [String: String] = [:],
-                                           then completion: @escaping ResultsHandler<ResourceType>) async
-        where ResourceType: Decodable {
-        let url = self.url(endpoint: endpoint, parameters: parameters)
-            await self.asyncFetchUrl(url,
-                          headers: headers,
-                          cachePolicy: cachePolicy,
-                          then: { (result: Result<Data, Error>, responseType: ResponseType) in
-            switch result {
-            case .success(let data):
-                do {
-                    let jsonParse = try self.jsonDecoder.decode(ResourceType.self, from: data)
-                    completion(Result.success(jsonParse), responseType)
-                } catch let error {
-                    completion(Result.failure(error), responseType)
-                }
-            case .failure(let error):
-                completion(Result.failure(error), responseType)
-            }
-        })
-    }
-
-    private func asyncFetchUrl(_ url: URL, headers:[String: String], cachePolicy: CachePolicy, then completion: @escaping ResultsHandler<Data>) async {
-        var dataTask: URLSessionDataTask?
-        var request = URLRequest(url: url)
-        for header in headers {
-            request.addValue(header.value, forHTTPHeaderField: header.key)
-        }
-        dataTask = await urlSession.dataTask(with: request,
-                                       completionHandler: { (data: Data?, response: URLResponse?, error: Error?) in
-            if let data = data {
-                if let response = response as? HTTPURLResponse {
-                    if response.statusCode != 200 {
-                        if cachePolicy == .networkElseCache,
-                            self.canFullfillRequestWithCache(request) {
-                            self.fullfillRequestWithCache(request, then: completion)
-                            return
-                        }
-                        completion(Result.failure(
-                            APIError.handleError(for: url,
-                                                 jsonDecoder: self.jsonDecoder,
-                                                 data: data,
-                                                 response: response)),
-                                   .network
-                        )
-                        return
-                    }
-                    let successMessage = "Success: 'GET' (\(response.statusCode)) \(url.absoluteString)"
-                    ContentstackLogger.log(.info, message: successMessage)
-
-                    CSURLCache.default.storeCachedResponse(
-                        CachedURLResponse(response: response,
-                                          data: data),
-                        for: request
-                    )
-                }
-                completion(Result.success(data), .network)
-                return
-            }
-
-            if let error = error {
-                if self.cachePolicy == .networkElseCache,
-                    self.canFullfillRequestWithCache(request) {
-                    self.fullfillRequestWithCache(request, then: completion)
-                    return
-                }
-                let errorMessage = """
-                Errored: 'GET' \(url.absoluteString)
-                Message: \(error.localizedDescription)
-                """
-                ContentstackLogger.log(.error, message: errorMessage)
-                completion(Result.failure(error), .network)
-                return
-            }
-
-        })
-        performDataTask(dataTask!, request: request, cachePolicy: cachePolicy, then: completion)
-    }
 }
+
 
 extension Stack {
